Fix clippy --all-features --all-targets -- -D warnings
Four lints from recent fixes:
  - redis_dedup.rs: manual_clamp — `.max(1).min(MAX)` → `clamp(1, MAX)`
  - bus.rs: unnecessary_cast — `stranded as u64` is u64→u64 (no-op
    after #59 widened in_flight_ingests to AtomicU64)
  - shard/mapper.rs: doc_lazy_continuation — paragraph break
    after the bullet list in the BUG #72 doc-comment
  - redis_dedup.rs: assertions_on_constants — moved
    `assert!(MAX_CAPACITY < usize::MAX, ...)` into a const block

All 1744 tests pass.

+++ net/crates/net/src/bus.rs
@@ -1227,7 +1227,7 @@ impl EventBus {
                 // surfaced-as-dropped contract.
                 self.stats
                     .events_dropped
-                    .fetch_add(stranded as u64, AtomicOrdering::Relaxed);
+                    .fetch_add(stranded, AtomicOrdering::Relaxed);
                 // BUG #48: also set the dedicated lossy-shutdown
                 // flag so `shutdown_via_ref` callers can detect a
                 // lossy outcome without parsing log lines or
